Add debug instrumentation to CompressedClassSpaceSizeInJmapHeap.java

+++ test/hotspot/jtreg/gc/metaspace/CompressedClassSpaceSizeInJmapHeap.java
@@ -46,6 +46,8 @@ import java.nio.charset.Charset;
 import java.util.List;
 
 public class CompressedClassSpaceSizeInJmapHeap {
+    static final KELVIN_DEBUG = true;
+
     // Note that on some platforms it may require root privileges to run this test.
     public static void main(String[] args) throws Exception {
         SATestUtils.skipIfCannotAttach(); // throws SkippedException if attach not expected to work.
@@ -67,23 +69,41 @@ public class CompressedClassSpaceSizeInJmapHeap {
         File err = new File("CompressedClassSpaceSizeInJmapHeap.stderr.txt");
         pb.redirectError(err);
 
-        // Sleep for 0.5s to make sure theApp is ready for attach from jhsdb
-        Thread.sleep(500);
-        run(pb);
+        // Sleep to give LingeredApp an opportunity to reach attachable status
+        Thread.sleep(1000);
+        int exitValue = run(pb);
 
-        OutputAnalyzer output = new OutputAnalyzer(read(out));
+	if (KELVIN_DEBUG) {
+            String stdout = read(out);
+            String stderr = read(err);
+	    System.out.println("");
+            System.out.println("Stdout is: ");
+            System.out.print(stdout);
+            System.out.println("");
+            System.out.println("Stderr is: ");
+            System.out.print(stderr);
+
+	    if (exitValue != 0) {
+                throw new Exception("jmap -heap exited with error code: " + exitValue);
+            }
+
+            OutputAnalyzer output = new OutputAnalyzer(stdout);
+	} else {
+            OutputAnalyzer output = new OutputAnalyzer(read(out));
+        }
         output.shouldContain("CompressedClassSpaceSize = 50331648 (48.0MB)");
         out.delete();
 
         LingeredApp.stopApp(theApp);
     }
 
-    private static void run(ProcessBuilder pb) throws Exception {
+    private static int run(ProcessBuilder pb) throws Exception {
         OutputAnalyzer output = ProcessTools.executeProcess(pb);
         int exitValue = output.getExitValue();
-        if (exitValue != 0) {
+        if (!KELVIN_DEBUG && (exitValue != 0)) {
             throw new Exception("jmap -heap exited with error code: " + exitValue);
         }
+        return exitValue;
     }
 
     private static String read(File f) throws Exception {
